Moved close method into class
