+++ src/instanceMethods.js
@@ -1,6 +1,5 @@
 export * from './instanceMethods/hideLoading.js'
 export * from './instanceMethods/getInput.js'
-export * from './instanceMethods/close.js'
 export * from './instanceMethods/enable-disable-elements.js'
 export * from './instanceMethods/show-reset-validation-error.js'
 export * from './instanceMethods/progress-steps.js'
